fix sass compile deprecated warning
use @else if instead @elseif and remove !global
Deprecation Warning on line 573, column 4 of assets/sass/libs/_skel.scss:
@elseif is deprecated and will not be supported in future Sass versions.
Use "@else if" instead.
    ╷
573 │             @elseif $expandValue {
    │             ^^^^^^^
    ╵

Deprecation Warning: As of Dart Sass 2.0.0, !global assignments won't be able to
declare new variables. Since this assignment is at the root of the stylesheet,
the !global flag is unnecessary and can safely be removed.

  ╷
7 │     $breakpoints: () !global;
  │     ^^^^^^^^^^^^^^^^^^^^^^^^
  ╵
    assets/sass/libs/_skel.scss 7:2  @import
    assets/sass/main.scss 4:9        root stylesheet